Change interaction of module loader and sources
This commit obsoletes the concept of ModuleID and introduces
ModuleSourceId instead. The idea is that the piece of information
that actually matters about a given module is its (full) name.
It is responsibility of frontend to map a full module name to
a source, which we identify by its source id (source ids are
source-storage specific). SourceInfoProvider is expected to
be able to map definitions to soures they were loaded from because
it makes sense to show where exactly the failing definition was
loaded from in case there are multiple places we get modules from.