Rework testing infrastructure
Kinda improves #34. Somewhat related to #36.

This simplifies and unifies tests of name resolver and type checker.
The idea is to store as much as possible in instance variables and
to keep everything in one place.

The one major negative consequence of this approach is that now
Prelude is being type checked before every single test. Hopefully, this
will be addressed later.